api_server/requirements.txt
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
connexion==2.7.0
Flask==1.1.2
Flask-AuditLog==1.0
Flask-Cors==3.0.10
Flask-No-Cache==1.0
Flask-SSLify==0.1.5
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-datastore==2.1.0
google-cloud-kms==2.2.0
google-cloud-secret-manager==2.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
libcst==0.3.17
MarkupSafe==1.1.1
mypy-extensions==0.4.3
openapi-schema-validator==0.1.4
openapi-spec-validator==0.3.0
packaging==20.9
proto-plus==1.14.2
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
swagger-ui-bundle==0.0.8
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug==0.16.1
